save and read csv

+++ Assets/Script/ReadCsv.cs
@@ -4,53 +4,56 @@ using UnityEngine;
 using UnityEditor;
 using UnityEngine.Networking;
 using AdvancedEditorTools.Attributes;
-using System;
+using System.IO;
 
 public class ReadCsv : MonoBehaviour
 {
     //[SerializeField] private NumberTowerGameManager m_NumberTowerGameManager;
     [SerializeField] private string m_LevelPath = "";
+    [SerializeField] private string m_ResourcesPath = "";
 #if UNITY_EDITOR
     //[MenuItem("Tool/ReadCSV")]
-    [Button("ReadCsv")]
-    private void ReadCsvFile(){
+    [Button("SaveCsv")]
+    private void SaveCsvFile(){
         StartCoroutine(GetCsvFromGoogle());
     }
-    private IEnumerator GetCsvFromGoogle(){
+    [Button("ReadCsv")]
+    private void ReadCsvFile(){
+
+        Dictionary<int,NumberTowerRoomData> allRooms = new Dictionary<int, NumberTowerRoomData>(); 
+
+        // room
+        if(Resources.Load<TextAsset>("Room") == null){
+            Debug.LogError(m_ResourcesPath+"/Room.csv is null" );
+            return;
+        }
+        string json = Resources.Load<TextAsset>("Room").ToString();
+
+        var contents = json.Split('\n',',');
+        for (int i = 5; i < contents.Length; i=i+5)
+        {
+            NumberTowerRoomData room = new NumberTowerRoomData();
+            room.Number = int.Parse(contents[i+1]);
+            Color bgColor = Color.black;
+            ColorUtility.TryParseHtmlString(contents[i+2], out bgColor);
+            room.BGColor = bgColor;
+            room.Operations = GetOperationByString(contents[i+3]);
+            string test = contents[i+4].Trim();
+            room.Prefab = GetNumberTowerPrefabByString(test);
+            allRooms.Add(int.Parse(contents[i].Trim()),room);
+        }
+
+        // Level
+        if(Resources.Load<TextAsset>("Level") == null){
+            Debug.LogError(m_ResourcesPath+"/Level.csv is null" );
+            return;
+        }
+        
         FileUtil.DeleteFileOrDirectory(m_LevelPath);
         AssetDatabase.Refresh();
 
-        // Room
-        Dictionary<int,NumberTowerRoomData> allRooms = new Dictionary<int, NumberTowerRoomData>(); 
-        UnityWebRequest www = UnityWebRequest.Get("https://docs.google.com/spreadsheets/d/e/2PACX-1vRfQPFboJJNlJ46EeqdMaTNR9GwYHOu5xiuIHuRm6dWsVvSQ10zE0c7ixeqxS857Y59TJ_OPNu_xymB/pub?gid=719668107&single=true&output=csv");
-        yield return www.SendWebRequest();
-        if(www.result == UnityWebRequest.Result.ConnectionError || www.result == UnityWebRequest.Result.ProtocolError){
-            Debug.Log("Error: " + www.error);
-        }else{
-            string json = www.downloadHandler.text;
-            var contents = json.Split('\n',',');
-            for (int i = 5; i < contents.Length; i=i+5)
-            {
-                NumberTowerRoomData room = new NumberTowerRoomData();
-                room.Number = int.Parse(contents[i+1]);
-                Color bgColor = Color.black;
-                ColorUtility.TryParseHtmlString(contents[i+2], out bgColor);
-                room.BGColor = bgColor;
-                room.Operations = GetOperationByString(contents[i+3]);
-                string test = contents[i+4].Trim();
-                room.Prefab = GetNumberTowerPrefabByString(test);
-                allRooms.Add(int.Parse(contents[i].Trim()),room);
-            }
-        }
-        
-        // level
-        www = UnityWebRequest.Get("https://docs.google.com/spreadsheets/d/e/2PACX-1vRfQPFboJJNlJ46EeqdMaTNR9GwYHOu5xiuIHuRm6dWsVvSQ10zE0c7ixeqxS857Y59TJ_OPNu_xymB/pub?output=csv");
-        yield return www.SendWebRequest();
-        if(www.result == UnityWebRequest.Result.ConnectionError || www.result == UnityWebRequest.Result.ProtocolError){
-            Debug.Log("Error: " + www.error);
-        }else{
-            string json = www.downloadHandler.text;
-            var contents = json.Split('\n',',');
+        json = Resources.Load<TextAsset>("Level").ToString();
+        contents = json.Split('\n',',');
             // level
             for (int i = 4; i < contents.Length; i=i+4)
             {
@@ -80,13 +83,47 @@ public class ReadCsv : MonoBehaviour
                 }
 
                 AssetDatabase.CreateAsset(levelScriptable, path);
-                //AssetDatabase.SaveAssets();
-                //EditorUtility.FocusProjectWindow();
             }
-        }
-
         
+        AssetDatabase.SaveAssets();
+        AssetDatabase.Refresh();
+    }
 
+
+
+
+
+    private IEnumerator GetCsvFromGoogle(){
+
+        // Room
+        UnityWebRequest www = UnityWebRequest.Get("https://docs.google.com/spreadsheets/d/e/2PACX-1vRfQPFboJJNlJ46EeqdMaTNR9GwYHOu5xiuIHuRm6dWsVvSQ10zE0c7ixeqxS857Y59TJ_OPNu_xymB/pub?gid=719668107&single=true&output=csv");
+        yield return www.SendWebRequest();
+        if(www.result == UnityWebRequest.Result.ConnectionError || www.result == UnityWebRequest.Result.ProtocolError){
+            Debug.Log("Error: " + www.error);
+        }else{
+            
+            FileUtil.DeleteFileOrDirectory(m_ResourcesPath+"/Room.csv");
+            AssetDatabase.SaveAssets();
+            AssetDatabase.Refresh();
+            string json = www.downloadHandler.text;
+            File.AppendAllText(m_ResourcesPath+"/Room.csv",json);
+
+            
+        }
+        
+        // level
+        www = UnityWebRequest.Get("https://docs.google.com/spreadsheets/d/e/2PACX-1vRfQPFboJJNlJ46EeqdMaTNR9GwYHOu5xiuIHuRm6dWsVvSQ10zE0c7ixeqxS857Y59TJ_OPNu_xymB/pub?output=csv");
+        yield return www.SendWebRequest();
+        if(www.result == UnityWebRequest.Result.ConnectionError || www.result == UnityWebRequest.Result.ProtocolError){
+            Debug.Log("Error: " + www.error);
+        }else{
+            FileUtil.DeleteFileOrDirectory(m_ResourcesPath+"/Level.csv");
+            AssetDatabase.SaveAssets();
+            AssetDatabase.Refresh();
+            string json = www.downloadHandler.text;
+            File.AppendAllText(m_ResourcesPath+"/Level.csv",json);
+
+        }
         
         AssetDatabase.SaveAssets();
         AssetDatabase.Refresh();
